Fixed screenshot folder errors.

+++ Continuum2018/Assets/Scripts/Camera/HiResScreenShots.cs
@@ -48,37 +48,57 @@ public class HiResScreenShots : MonoBehaviour
 
 		if (takeHiResShot == true) 
 		{
-			RenderTexture rt = new RenderTexture (resWidth, resHeight, 24);
-			cam.targetTexture = rt;
-			Texture2D screenShot = new Texture2D (resWidth, resHeight, TextureFormat.RGB24, false);
-			cam.Render();
-			RenderTexture.active = rt;
-			screenShot.ReadPixels (new Rect (0, 0, resWidth, resHeight), 0, 0);
-			cam.targetTexture = null;
-			RenderTexture.active = null; // JC: added to avoid errors
-			Destroy(rt);
-			byte[] bytes = screenShot.EncodeToPNG();
-			string filename = ScreenShotName (resWidth, resHeight);
-			System.IO.File.WriteAllBytes (filename, bytes);
-			Debug.Log (string.Format("Took screenshot to: {0}", filename));
+			if (System.IO.Directory.Exists (Application.dataPath + "/" + "Screenshots") == false)
+			{
+				System.IO.Directory.CreateDirectory (Application.dataPath + "Screenshots");
+				Debug.Log ("Screenshot folder was missing, created new one.");
+			}
+				
+			if (System.IO.Directory.Exists (Application.dataPath + "/" + "Screenshots") == true)
+			{
+				RenderTexture rt = new RenderTexture (resWidth, resHeight, 24);
+				cam.targetTexture = rt;
+				Texture2D screenShot = new Texture2D (resWidth, resHeight, TextureFormat.RGB24, false);
+				cam.Render();
+				RenderTexture.active = rt;
+				screenShot.ReadPixels (new Rect (0, 0, resWidth, resHeight), 0, 0);
+				cam.targetTexture = null;
+				RenderTexture.active = null; // JC: added to avoid errors
+				Destroy(rt);
+				byte[] bytes = screenShot.EncodeToPNG();
+				string filename = ScreenShotName (resWidth, resHeight);
+				System.IO.File.WriteAllBytes (filename, bytes);
+				Debug.Log (string.Format("Took screenshot to: {0}", filename));
+			}
+				
 			takeHiResShot = false;
 		}
 
 		if (takeHiDoubleResShot == true) 
 		{
-			RenderTexture rt = new RenderTexture (resWidth * 2, resHeight * 2, 24);
-			cam.targetTexture = rt;
-			Texture2D screenShot = new Texture2D (resWidth* 2, resHeight* 2, TextureFormat.RGB24, false);
-			cam.Render();
-			RenderTexture.active = rt;
-			screenShot.ReadPixels (new Rect (0, 0, resWidth * 2, resHeight * 2), 0, 0);
-			cam.targetTexture = null;
-			RenderTexture.active = null; // JC: added to avoid errors
-			Destroy(rt);
-			byte[] bytes = screenShot.EncodeToPNG();
-			string filename = ScreenShotName (resWidth * 2, resHeight * 2);
-			System.IO.File.WriteAllBytes (filename, bytes);
-			Debug.Log (string.Format("Took screenshot to: {0}", filename));
+			if (System.IO.Directory.Exists (Application.dataPath + "/" + "Screenshots") == false) 
+			{
+				System.IO.Directory.CreateDirectory (Application.dataPath + "/" + "Screenshots");
+				Debug.Log ("Screenshot folder was missing, created new one.");
+			}
+
+			if (System.IO.Directory.Exists (Application.dataPath + "/" + "Screenshots") == true) 
+			{
+				RenderTexture rt = new RenderTexture (resWidth * 2, resHeight * 2, 24);
+				cam.targetTexture = rt;
+				Texture2D screenShot = new Texture2D (resWidth* 2, resHeight* 2, TextureFormat.RGB24, false);
+				cam.Render();
+				RenderTexture.active = rt;
+				screenShot.ReadPixels (new Rect (0, 0, resWidth * 2, resHeight * 2), 0, 0);
+				cam.targetTexture = null;
+				RenderTexture.active = null; // JC: added to avoid errors
+				Destroy(rt);
+				byte[] bytes = screenShot.EncodeToPNG();
+				string filename = ScreenShotName (resWidth * 2, resHeight * 2);
+				System.IO.File.WriteAllBytes (filename, bytes);			
+				Debug.Log (string.Format("Took screenshot to: {0}", filename));
+			}
+				
 			takeHiDoubleResShot = false;
 		}
 	}
